Implemented most of the CPU operations.

+++ emu/src/main/java/joelbits/emu/Memory.java
@@ -1,7 +1,7 @@
 package joelbits.emu;
 
 public interface Memory {
-	short readFromMemory(int location);
-	void writeToMemory(short data, int location);
+	int readFromMemory(int location);
+	void writeToMemory(int data, int location);
 	void clearMemory();
 }
